Spit stream conformance tets - Fix #1921
